feat(studio): move Execute and AI Generate to floating canvas buttons
Move primary action buttons from the toolbar to floating positions
directly on the canvas for better spatial proximity to their results:

- Execute button: bottom-right of canvas (was toolbar right)
- AI Generate button: right edge near AI panel (was toolbar left)

This addresses UX issues where action buttons were spatially
disconnected from their results (1,131px gap for AI Generate,
57px vertical gap for Execute).

Also consolidates node header accent line CSS into a base rule
with `width: auto !important` to prevent fixed-width issues on
wider nodes.

Changes:
- WorkflowEditor: Add floating action buttons container
- Toolbar: Remove Execute and AI Generate buttons
- App: Remove unused onToggleAI/isAIPanelOpen props from Toolbar
- CSS: Add .floimg-canvas-actions styles, consolidate ::before rules
